+++ query-result-writer.hh
@@ -50,7 +50,7 @@ class result::partition_writer {
     ser::vector_position _pos;
     digester& _digest;
     digester _digest_pos;
-    uint32_t& _row_count;
+    uint64_t& _row_count;
     uint32_t& _partition_count;
     api::timestamp_type& _last_modified;
 public:
@@ -62,7 +62,7 @@ public:
         ser::vector_position pos,
         ser::after_qr_partition__key<bytes_ostream> w,
         digester& digest,
-        uint32_t& row_count,
+        uint64_t& row_count,
         uint32_t& partition_count,
         api::timestamp_type& last_modified)
         : _request(request)
@@ -107,7 +107,7 @@ public:
     digester& digest() {
         return _digest;
     }
-    uint32_t& row_count() {
+    uint64_t& row_count() {
         return _row_count;
     }
     uint32_t& partition_count() {
@@ -124,7 +124,7 @@ class result::builder {
     const partition_slice& _slice;
     ser::query_result__partitions<bytes_ostream> _w;
     result_request _request;
-    uint32_t _row_count = 0;
+    uint64_t _row_count = 0;
     uint32_t _partition_count = 0;
     api::timestamp_type _last_modified = api::missing_timestamp;
     short_read _short_read;
@@ -147,7 +147,7 @@ public:
 
     const partition_slice& slice() const { return _slice; }
 
-    uint32_t row_count() const {
+    uint64_t row_count() const {
         return _row_count;
     }
